Refactor function to extract mappings as a hook
